+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 7.4
 - Update French & Spanish translations in rosariosis.po, help.po
 - Add Total Credits in ReportCards.fnc.php
 - Do not display "General Comments" title if no comments in ReportCards.fnc.php
+- HTML display rows of 3 School Period checkboxes in AddAbsences.php
 
 Changes in 7.3.1
 ----------------