+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 7.5
+--------------
+- HTML fix Student Assignment Submission display in StudentAssignments.fnc.php
+
 Changes in 7.4
 --------------
 - List sort comment: trim & fix position in ListOutput.fnc.php
@@ -19,7 +23,6 @@ Changes in 7.4
 - Fix SQL error integer out of range in Food_Service/Students/Accounts.php
 - French translation: replace "Effacer" with "Supprimer" in rosariosis.po & help.po
 - Fix Teacher Programs Progress Report PDF, do not echo form in TeacherPrograms.php
-- HTML fix Student Assignment Submission display in StudentAssignments.fnc.php
 
 Changes in 7.3.1
 ----------------